test: split test cases out to share with ts
